refactor(reply): canonicalize reply context module
