+++ local_server.py
@@ -101,7 +101,7 @@ from runtime.memory.private_world_runtime import (
     create_private_world_runtime,
     resolve_private_world_database,
 )
-from reply_context import (
+from runtime.reply.reply_context import (
     ReplyContext,
     ReplyMode,
     TrustedTime,
