+++ runtime/reply/reply_quality_gate.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from enum import StrEnum
 from typing import Any, Mapping, Protocol, Sequence
 
-from reply_context import ReplyContext
+from runtime.reply.reply_context import ReplyContext
 from runtime.reply.reply_policy import scan_reply
 from runtime.reply.reply_reviewer import ReviewResult, ReviewStatus, ReviewVerdict
