+++ tests/persona/test_reply_context.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 from jsonschema import Draft202012Validator
 
-from reply_context import (
+from runtime.reply.reply_context import (
     ReplyContext,
     ReplyContextError,
     BehaviorLevel,
