+++ tests/persona/test_reply_policy.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from reply_context import (
+from runtime.reply.reply_context import (
     OutputChannel,
     OutputConstraints,
     ReplyContext,
