+++ tests/persona/test_reply_quality_gate.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from reply_context import ReplyContext, ReplyMode, TrustedTime
+from runtime.reply.reply_context import ReplyContext, ReplyMode, TrustedTime
 from runtime.reply.reply_quality_gate import QualityGateStatus, run_reply_quality_gate
 from runtime.reply.reply_reviewer import (
     NullReviewer,
